Playing with play framework
Setting up routes for testing controllers and views.

Made a stupid simple view for fooling with templates in scala and
learning about that, then defined the routes in application.conf to
be the routes file (something tells me explicitly doing this saves a few
nanoseconds)

Made a Review.scala file but put it in the wrong place but it's a
controller I swear! Making a sample JSON writer that uses implicit
traits and fooling with it.